+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+### Added
+- Add `EnvelopeDocumentsEndpoint.merge`
 
 ## [1.4.0] - 2021-09-17
 ### Added