+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeProperties` resource and `Envelope.properties`
 - Add `EnvelopeTemplateTag.assignment` + `EnvelopeTemplateTag.EnvelopeRecipient` can be null
 - Add `EnvelopeTag.assignment` + `EnvelopeTag.EnvelopeRecipient` can be null
+- Add `EnvelopeTemplate.useDefaultTemplateSettings`
 
 ## [1.4.0] - 2021-09-17
 ### Added