+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeTemplate.useDefaultTemplateSettings`
 - Add `EnvelopeDocument.labelPositioning`, `EnvelopeDocument.labelPositionX`, `EnvelopeDocument.labelPositionY`, 
 - Add `EnvelopeTemplateDocument.labelPositioning`, `EnvelopeTemplateDocument.labelPositionX`, `EnvelopeTemplateDocument.labelPositionY`, 
+- Add `AccountSettings.bankIdSign`
 
 ## [1.4.0] - 2021-09-17
 ### Added