+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [1.5.0] - 2021-11-09
 ### Added
 - Add `EnvelopeDocumentsEndpoint.merge`
 - Add `EnvelopeTemplatesEndpoint.clone`